+++ srbuild/__init__.py
@@ -1,1 +1,3 @@
-from srbuild.tools import compiler, linker
+from srbuild.project.project import Project
+from srbuild.cli.cli import cli
+from srbuild.logger import G_LOGGER