+++ srbuild/generator/__init__.py
@@ -1,0 +1,1 @@
+from srbuild.generator.rbuild import RBuildGenerator