+++ srbuild/project/target.py
@@ -3,4 +3,6 @@ from typing import NewType, Dict
 
 # A Dict[str, LinkedNode] that maps profile names to the LinkedNodes for a target
 class ProjectTarget(dict):
-    pass
+    def __init__(self, name, *args, **kwargs):
+        super().__init__(self, *args, **kwargs)
+        self.name = name